requirements.txt
ppci
nmigen
numpy
pytest
pytest-xdist
pyelftools
